Added param for doing running the script

+++ universal_schema/bin/learning/run_convert2json.sh
@@ -1,8 +1,9 @@
 #!/usr/bin/env bash
 echo "Starting the convert2json script"
 # Parse command line args.
-while getopts ":r:t:d:" opt; do
+while getopts ":p:r:t:d:" opt; do
     case "$opt" in
+        p) parse_type=$OPTARG ;;
         r) relationship_id_file=$OPTARG ;;
         t) train_split=$OPTARG ;;
         d) dev_split=$OPTARG ;;
@@ -21,20 +22,24 @@ source_path="$CUR_PROJ_DIR/src/learning"
 data_source_path="$CUR_PROJ_DIR/datasets_proc/freebase/latfeatus"
 script_name="convert2json"
 
-# Shuffling the input - relationship to entity file
-cd $data_source_path
-shuf "$relationship_id_file" > "$relationship_id_file-shuf"
-echo "Created: $relationship_id_file-shuf"
-num_of_lines=$(< "$relationship_id_file" wc -l)
+if [[ $parse_type == 'split' ]]; then
+    # Shuffling the input - relationship to entity file
+    cd $data_source_path
+    shuf "$relationship_id_file" > "$relationship_id_file-shuf"
+    echo "Created: $relationship_id_file-shuf"
+    num_of_lines=$(< "$relationship_id_file" wc -l)
 
-# Clean the stale train,dev and test files
-rm -f train*
-rm -f dev*
-rm -f test*
-echo "Cleaned the dataset directory"
+    # Clean the stale train,dev and test files
+    rm -f train*
+    rm -f dev*
+    rm -f test*
+    echo "Cleaned the dataset directory"
+fi
+
 
 cmd="python2 $source_path/$script_name.py
     $action \
+    --parse_type $parse_type\
     --relationship_id_file $relationship_id_file \
     --train_split $train_split \
     --dev_split $dev_split \
